Improve docs for kubelet
