+++ crates/kubelet/src/image_client.rs
@@ -1,9 +1,12 @@
+//! Container module image clients that can fetch images from remote stores
 use async_trait::async_trait;
 
 use oci_distribution::Reference;
 
+/// An image client capable of fetching images from a remote store
 #[async_trait]
 pub trait ImageClient {
+    /// Given a certain image reference pull the image data from the remote store
     async fn pull(&mut self, image: &Reference) -> anyhow::Result<Vec<u8>>;
 }
